requirements.txt
# Numbers & Data
numpy==1.24.3
# pandas>=2.0.1
geopandas==0.12.2

# User Interface
streamlit==1.21.0
plotly==5.14.1

# Machine Learning
scikit-learn==1.2.2